+++ app/src/main/java/org/android/go/sopt/repository/ListUsersRepository.kt
@@ -1,0 +1,8 @@
+package org.android.go.sopt.repository
+
+import org.android.go.sopt.data.remote.model.ResponseListUsersDto
+import retrofit2.Response
+
+interface ListUsersRepository {
+    suspend fun getListUsers(): Response<ResponseListUsersDto>
+}
